Allow adding node for relational question without contact chooser.
User can search for existing nodes or create new node. The node will be
added to the "availableGroups" of the current question, so that the
newly added node will show up in this question for every user. 

Survey designer need to be carefully arrange the group
relationships because this functionality allow user to add node to
group(s) where in normal case impossible. For example, user can create a
node of type "Document" and add this node to the question whose
available group is "User". In normal situation, a node of type
"Document" should not be able to added into group "User".